Write a new README and add a license file.
Based on the source code history, it seems that significant portions
of the original plugin were copied from active_scaffold, so I'm
concluding that the code is under the same MIT license.  There are no
copyright marks in the history, but if you're a copyright holder who
objects to your code being under the MIT license, please let me know.